+++ packages/react-magma-dom/src/components/TimePicker/index.tsx
@@ -105,7 +105,8 @@ const StyledLegend = styled.legend<{
 
 const InputsContainer = styled.div<{ theme: ThemeInterface }>`
   background: ${props => props.theme.colors.neutral08};
-  left: ${props => props.theme.spaceScale.spacing09};
+  left: ${props => props.theme.spaceScale.spacing08};
+  margin-left: ${props => props.theme.spaceScale.spacing02};
   position: absolute;
   top: ${props => props.theme.spaceScale.spacing03};
 `;