+++ packages/react-magma-dom/src/theme/magma.ts
@@ -33,6 +33,7 @@ export interface Colors {
   tint: string;
   tint02: string;
   tint03: string;
+  tint04: string;
   tone: string;
   tone02: string;
 }
@@ -46,6 +47,7 @@ export interface Breakpoints {
 }
 
 export interface IconSizes {
+  xSmall: number;
   small: number;
   medium: number;
   large: number;
@@ -317,9 +319,10 @@ export const magma = {
   },
 
   iconSizes: {
-    small: 16,
-    medium: 20,
-    large: 28,
+    xSmall: 16,
+    small: 20,
+    medium: 24,
+    large: 32,
   },
 
   // Typography